move URI to ssi-core
