+++ ssi-core/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+thiserror = "1.0"
 async-trait = "0.1"
 serde = { version = "1.0", features = ["derive"] }
